requirements.txt
aiogram==3.7.0
anthropic==0.25.0
SQLAlchemy==2.0.30
aiomysql==0.2.0
alembic==1.13.1
openpyxl==3.1.2
APScheduler==3.10.4
python-dotenv==1.0.1
Pillow==10.3.0
aiohttp==3.9.5
